+++ src/App.css
@@ -26,3 +26,18 @@
 .mb {
   margin-bottom: 2rem;
 }
+
+.clear-button {
+  display: flex;
+  justify-content: end;
+}
+
+.clear-button > button {
+  background: none;
+  color: inherit;
+  border: none;
+  padding: 0;
+  font: inherit;
+  cursor: pointer;
+  outline: inherit;
+}
