+++ src/components/index.js
@@ -1,0 +1,2 @@
+export { default as TodoList } from "./TodoList";
+export { default as TodoToggleButtons } from "./TodoToggleButtons";
